README FIRST
Decided to start this off by documenting first and coding second. This
will be the frist post on the main branch. Development work to meet
the readme requirements will take place in a seperate dev branch and
updates to the main branch until the code is ready will only be to the
readme file.

Thats the plan at any rate.